requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosmtplib==2.0.2
aiosqlite==0.19.0
alembic==1.11.1
anyio==3.7.0
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
bbcode==1.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
chevron==0.14.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
decorator==5.1.1
emoji==2.5.0
exceptiongroup==1.1.1
fastapi==0.96.1
filelock==3.12.1
fonttools==4.39.4
frozenlist==1.3.3
fsspec==2023.6.0
greenlet==2.0.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httptools==0.5.0
huggingface-hub==0.15.1
hyperframe==6.0.1
idna==3.4
importlib-metadata==4.13.0
jieba==0.42.1
Jinja2==3.1.2
kiwisolver==1.4.4
loguru==0.7.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
msgpack==1.0.5
multidict==6.0.4
nanoid==2.0.0
nonebot-adapter-console==0.3.2
nonebot-adapter-onebot==2.2.3
nonebot-plugin-apscheduler==0.2.0
nonebot-plugin-chatrecorder==0.2.3
nonebot-plugin-datastore==0.6.3
nonebot-plugin-gocqhttp==0.6.10
nonebot-plugin-htmlrender==0.2.0.3
nonebot-plugin-localstore==0.4.1
nonebot-plugin-picstatus==0.5.0
nonebot-plugin-send-anything-anywhere==0.2.7
nonebot-plugin-wordcloud==0.4.8
nonebot2==2.0.0
numpy==1.24.3
openai==0.27.8
opencv-python-headless==4.7.0.72
packaging==23.1
pil-utils==0.1.7
Pillow==9.5.0
playwright==1.34.0
psutil==5.9.5
py-cpuinfo==9.0.0
pydantic==1.10.9
pyee==9.0.4
Pygments==2.15.1
pygtrie==2.5.0
pymdown-extensions==10.0.1
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-markdown-math==0.8
pytz==2023.3
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rfc3986==1.5.0
rich==13.4.1
safetensors==0.3.1
self==2020.12.3
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.16
starlette==0.27.0
StrEnum==0.4.10
textual==0.10.1
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.65.0
transformers==4.30.1
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
win32-setctime==1.1.0
wordcloud==1.9.2
yarl==1.9.2
zipp==3.15.0

httpx~=0.23.3